kafkaConsumer/requirements.txt
confluent-kafka==2.8.2
python-dotenv==1.0.1
